Fixed minor bug in subtitle format FcpXml
git-svn-id: https://subtitleedit.googlecode.com/svn/trunk@335 99eadd0c-20b8-1223-b5c4-2a2b2df33de2

+++ src/Logic/SubtitleFormats/FinalCutProXml.cs
@@ -106,7 +106,7 @@ namespace Nikse.SubtitleEdit.Logic.SubtitleFormats
 
         public override void LoadSubtitle(Subtitle subtitle, List<string> lines, string fileName)
         {
-            const int frameRate = 30;
+            const double frameRate = 30.0d;
             _errorCount = 0;
 
             StringBuilder sb = new StringBuilder();
